Move the choosing of the reporter in to `reporter.js`.
