+++ lib/kavun.js
@@ -5,15 +5,12 @@ import { startTimer } from './utils/time-tracker';
 import { runner } from './runner';
 import { unitCollector } from './index';
 import { loadTestFiles } from './test-file-loader';
-import { reporter as consoleReporter } from './reporters/console-reporter';
+import { createReporter } from './reporter.js';
 
 const path = process.argv[2];
 const reporterName = process.argv[3] || 'console';
 
-const reporters = {
-  'console': consoleReporter,
-};
-const config = { reporter: reporters[reporterName] };
+const config = { reporter: createReporter(reporterName) };
 const stopTimer = startTimer();
 loadTestFiles(require, join(process.cwd(), path));
 runner(config, { unitCollector, stopTimer });
