+++ lib/reporter.js
@@ -1,0 +1,5 @@
+import { reporter as consoleReporter } from './reporters/console-reporter.js';
+
+export const createReporter = (name) => {
+  return consoleReporter;
+};
