--- lib/reporters/index.js
@@ -1,24 +1,0 @@
-export const reporter = (outputDevice) => {
-  const successSign = '✓';
-  const failureSign = 'x';
-
-  const log = (message) => outputDevice.log(message);
-
-  const step = (description, result, elapsedTime) => {
-    const sign = result ? successSign : failureSign;
-
-    log(`${description} => ${sign} in ${elapsedTime.toFixed(2)}ms`);
-  };
-
-  const result = (failed, succeeded, elapsedTime) =>
-    log(`${failed} failed, ${succeeded} succeeded in ${elapsedTime.toFixed(2)}ms`);
-
-  const newLine = () => log('\n');
-
-  return {
-    step,
-    result,
-    log,
-    newLine
-  }
-};
